COMMAND LINE PARAMETERS
Changed and added command line parameters. Previously, the whole url was split into parts: protocol name, server name, port. This is due to the need to set the port at the start of the application, so that on one machine it would be possible to run several stations, dividing them by ports