DOWN VERSION
Made the version number less than 1. Because it doesn't work as expected yet.
Removed sending messages from UseCase diagram - this should be done by scripts.
In UseCase diagram made a requirement to create a secure agent. It is through him that the code will be run on this machine or on another machine